requirements.txt
bcrypt==3.2.0
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
dnspython==2.1.0
ecdsa==0.14.1
email-validator==1.1.2
fastapi==0.63.0
google-api-core==1.26.0
google-auth==1.27.0
google-cloud-core==1.6.0
google-cloud-storage==1.36.1
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
h11==0.12.0
httpcore==0.12.3
httpx==0.17.0
idna==2.10
mongoengine==0.22.1
Naked==0.1.31
packaging==20.9
passlib==1.7.4
protobuf==3.15.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.10.1
pycryptodomex==3.10.1
pydantic==1.8
pymongo==3.11.3
pyparsing==2.4.7
python-dotenv==0.15.0
python-jose==3.2.0
python-multipart==0.0.5
pytz==2021.1
PyYAML==5.4.1
rave-python==1.2.10
requests==2.25.1
rfc3986==1.4.0
rsa==4.7.2
shellescape==3.8.1
six==1.15.0
sniffio==1.2.0
starlette==0.13.6
typing-extensions==3.7.4.3
urllib3==1.26.3
uvicorn==0.13.4
